Made the first things
I tried to solve a part of our exercise on my own, but somehow it
escalated... I didn't want to write that much code...
What it does ATM:
Get a therm via an input window.
Analyze the therm.
Get get part we want to solve.
Get an agent, who could solve our problem and ask for his service.
The agent itself is able to tell the other, that he can solve the job.
The the consumerAgent sends him the numbers and the arithmeticAgent
analyzes the message and does his job.
At the end, the arithmeticAgent returns the number and the consumerAgent
saves the result.

However... ATM it only works with "*" and just with numbers in the therm
between 0-9...
Also it's quite buggy (;D)... Maybe you cold take a closer look at the
code.. I commented it a lot, i think.