Spaghetti Code V2
Completed what I think was the most work of my life :’D
Our program now correctly solves the given term in our task.
More testing to follow tomorrow and the last requirements have to be
fulfilled.

The program has a lot of redundant code in it (so I guess) but this is
no time for a polished sapphire!